LLM/requirements.txt
requests
playsound
python-dotenv
Flask
langchain
openai
pydub
pyaudio